[FIX] account@saas~12.3: remove account.view_account_bank_journal_form view
The view `account.view_account_bank_journal_form` exists in Odoo from `9.0`
to `13.0` but temporary disappears in `saas-12.3`. After discussion with the accounting team,
it appears that this view came back in `13.0` but as dead code (this view is not used).
So, this view will probably disappear again in `14.0`.
So, the best solution is simply to remove this view.

Upgrade request: 45307

closes odoo/upgrade#1370

Signed-off-by: Christophe Simonis <chs@odoo.com>