[FIX] account@saas~12.4: set `account_move` `currency_id` not null
Request 48191

In 12.0, the `currency_id` field on `account.move` is a compute field.
https://github.com/odoo/odoo/blob/2d45844315656e4ea5c9008fc2fd800466210d82/addons/account/models/account_move.py#L102
Given the implementation of the compute, this is almost impossible
to have a move with not currency set:
https://github.com/odoo/odoo/blob/2d45844315656e4ea5c9008fc2fd800466210d82/addons/account/models/account_move.py#L75-L76
In 13.0, this field become a regular, but *required*, field.
https://github.com/odoo/odoo/blob/13.0/addons/account/models/account_move.py#L127

To avoid the ORM to init the default values for this column
when adding the `NOT NULL` on the column,
set it in an upgrade script.

This is mainly to avoid the below issue to happen:
```
  File "/src/odoo/13.0/odoo/modules/registry.py", line 350, in init_models
    model._auto_init()
  File "/src/odoo/13.0/odoo/models.py", line 2482, in _auto_init
    new = field.update_db(self, columns)
  File "/src/odoo/13.0/odoo/fields.py", line 2430, in update_db
    return super(Many2one, self).update_db(model, columns)
  File "/src/odoo/13.0/odoo/fields.py", line 853, in update_db
    self.update_db_notnull(model, column)
  File "/src/odoo/13.0/odoo/fields.py", line 893, in update_db_notnull
    model._init_column(self.name)
  File "/src/odoo/13.0/odoo/models.py", line 2408, in _init_column
    value = field.default(self)
  File "/src/odoo/13.0/addons/account/models/account_move.py", line 84, in _get_default_currency
    journal = self._get_default_journal()
  File "/src/odoo/13.0/addons/account/models/account_move.py", line 74, in _get_default_journal
    raise UserError(error_msg)
odoo.exceptions.UserError: ('Please define an accounting miscellaneous journal in your company', '')
```

In addition, there might be a slight performance gain,
as the ORM won't try to load uselessly the defaults.

closes odoo/upgrade#1316

Signed-off-by: Denis Ledoux (dle) <dle@odoo.com>