[FIX] account@saas~12.4: rebind FK from invoice to move models
Requests 46508 46168 45683 45436 44545

When the invoice model is deleted,
the report `account.account_invoices` was deleted
because of the `ON DELETE CASCADE` on `binding_model_id`
https://github.com/odoo/odoo/blob/ff37268638b281207802473acf90f47c7b880ef2/odoo/addons/base/models/ir_actions.py#L45

And when `l10n_mx_edi` was being updated,
and it tried to update the report `account.account_invoices`,
and an Exception occured when attempting to update the report which was no longer there.

This issue can be solved by rebinding the records refenrencing the
invoice models to the move models, which is was is doing `merge_model`.

`merge_model` has been adapted to allow to not drop the table.

```
  File "/src/odoo/13.0/odoo/tools/convert.py", line 803, in convert_xml_import
    obj.parse(doc.getroot())
  File "/src/odoo/13.0/odoo/tools/convert.py", line 721, in parse
    exc_info[2]
  File "/src/odoo/13.0/odoo/tools/pycompat.py", line 13, in reraise
    raise value.with_traceback(tb)
  File "/src/odoo/13.0/odoo/tools/convert.py", line 712, in parse
    self._tag_root(de)
  File "/src/odoo/13.0/odoo/tools/convert.py", line 674, in _tag_root
    f(rec)
  File "/src/odoo/13.0/odoo/tools/convert.py", line 527, in _tag_record
    raise Exception("Cannot update missing record %r" % xid)
odoo.tools.convert.ParseError: "Cannot update missing record 'account.account_invoices'" while parsing /src/enterprise/13.0/l10n_mx_edi/views/l10n_mx_edi_report_invoice.xml:2, near
```

closes odoo/upgrade#1320

Signed-off-by: Christophe Simonis <chs@odoo.com>