[IMP] point_of_sale@saas~13.4: responsive UI
The point of sale user interface has been adapted to work on mobile
devices.

We modified the design of payment lines to look like payment methods and
stack them together in the mobile view.

When you had a cash method, the amount was not preffilled during
payment, now we always prefill the amount to keep it more simple.

TASK-ID: 1887609

See
https://github.com/odoo/odoo/pull/51891

closes odoo/upgrade#1334

Related: odoo/odoo#51891
Signed-off-by: Nicolas Seinlet (nse) <nse@odoo.com>