[FIX] account@saas~12.4: zero-line with price * qty * disc not zero
Force quantity to 0 to force the subtotal to be computed to 0
In some cases, the subtotal on the invoice line is 0, but price unit * quantity * discount is not,
and the override of the move line `create` will recompute the subtotal according to these.
https://github.com/odoo/odoo/blob/c5cd7329529cfa27d1f65217bd54042642711048/addons/account/models/account_move.py#L3307
It doesn't matter setting the quantity to 0 here,
because it's reset with the correct quantity at the end of this method
with a raw `UPDATE` sql on `account_move_line`, updating the quantity according to the invoice line.

Request 48600

closes odoo/upgrade#1348

Signed-off-by: Denis Ledoux (dle) <dle@odoo.com>