[IMP] account@saas~12.4: invoice line account company different from invoice company
The company of an invoice line account must be the same than the parent invoice company.
If it's not the case, the invoice line must be associated to another account.

The PR #1248 fixed this issue but in fact, a better solution consists in letting the ORM
computing itself the right account to use.

Upgrade request: 46849, 45579

closes odoo/upgrade#1356

Signed-off-by: Denis Ledoux (dle) <dle@odoo.com>