[FIX] account@saas~12.4: exclude from invoice move lines that were not in invoice lines
This commit introduces a change of paradigm.
Before:
 - `exclude_from_invoice_tab` is `NULL` by default (a.k.a Falsy)
 - Then, the scripts set them to `true` based on criteria on the move line itself
   e.g.
   - move lines with invoice types and a tax line or a receivable/payable account type
   ```
        UPDATE account_move_line aml
           SET exclude_from_invoice_tab = 't'
          FROM account_move am
         WHERE am.type in ('in_invoice', 'out_invoice', 'in_refund', 'out_refund', 'out_receipt', 'in_receipt')
           AND aml.move_id = am.id
           AND (aml.tax_line_id IS NOT NULL OR aml.account_internal_type IN ('receivable', 'payable'))
   ```
   - move lines with a zero amount added during the upgrade
   ```
        UPDATE account_move_line aml
           SET exclude_from_invoice_tab = true
          FROM account_move move
         WHERE move.id = aml.move_id
           AND aml.balance = 0
           AND aml.exclude_from_invoice_tab IS NULL
           AND move.type != 'entry'
   ```
The problem with the above is that some users added in move lines of their invoices
lines with accounts which are not receivable/payable, not even reconciliable, and should not be.
And those then appears in the invoice while it was not the case before.

After:
 - `exclude_from_invoice_tab` is `true` by default,
 - then, we set to `False` all move lines we know are invoice lines,
   thanks to the mapping
   `invl_aml_mapping`
   This is actually already done in the upgrade scripts:
   ```
   UPDATE account_move_line aml
      SET exclude_from_invoice_tab = FALSE
     FROM invl_aml_mapping m
    WHERE aml.id = m.aml_id
   ```

With this way, we are sure only lines that were invoice lines before upgrade
appears in the invoices.

Nevertheless, one thing to pay attention to:
`exclude_from_invoice_tab` is used by the temporary table `TEMPORARY TABLE am_untaxed`
to compute the total of the invoice. So it must be correctly set before creating/filling this table.

This is currently the case, the above mentioned UPDATE is done just before.

Besides that, in the upgrade script at least,
I see no other use of the `account_move_line` `exclude_from_invoice_tab` column

There is also `account/saas~12.4.1.1/post-00-fix-invoice-accounts.py`
to take into account,
but currently is completely useless because of the double call to `cr.fetchall`:
https://github.com/odoo/upgrade/blob/4ebec02f14d68ec1c46336069802ca50df70c63a/migrations/account/saas~12.4.1.1/post-00-fix-invoice-accounts.py#L29-L32
The second `cr.fetchall` will always return an empty list,
because the request was consumed by the above `res = cr.dictfetchall()`
Besides,
1. changing the `user_type_id` of the account seems a bit intrusive to me,
2. It doesn't work the cases/requests observed for this revision,
   because the problematic lines do not use the account of the invoice,
   and therefore the condition `AND inv.account_id = aml.account_id` filters them out.

Requests 49036, 48329

closes odoo/upgrade#1364

Signed-off-by: Denis Ledoux (dle) <dle@odoo.com>