[FIX] l10n_nl@saas~11.5: empty intrastat report
The tax `Verkopen export binnen EU` is split into two taxes
(Products and Services)
in odoo/odoo@02ccb58401a45528adb77c768318c9f6dfdf05b6

The related tag `tag_nl_13` is then split into two as well
`tag_nl_40` and `tag_nl_41`.

The intrastat report (ICP) is using these tags:
https://github.com/odoo/enterprise/blob/12.0/l10n_nl_reports/models/l10n_nl_report_intrastat.py#L53-L54

Without this script,
1. the former tax `Verkopen export binnen EU`
   completely loose its tag (as `tag_nl_13` is deleted)
2. the intrastat report is empty, as there is no tax using the tag `tag_nl_40`, `tag_nl_41`.

As before, there was only one tax, we can consider the tag `tag_nl_13`
is replaced by the tag `tag_nl_40`.
If the customer would like to have the two taxes,
he will have to create the second tax and assign the tag `tag_nl_41` to it,
and start using it in his invoices.

But for the past, this is too late: in the invoices,
the same tax has been used whatever it was a product or a service.

closes odoo/upgrade#1382

Signed-off-by: Denis Ledoux (dle) <dle@odoo.com>