[FIX] account@saas~12.4: give a chance to `check_amount_currency_balance_sign`
Fix move lines with the currency set to the company currency and no amount currency
to have better chances to be able to set the constraint `check_amount_currency_balance_sign`
https://github.com/odoo/odoo/blob/57277257efc62f0f58945797daf9a7d667991dde/addons/account/models/account_move.py#L2617-L2633

There are a bunch of databases having a currency set equals to the company currency on their move line,
because before it wasn't that enforced.

And if we don't give a chance to this SQL constraint to be applied,
it will still not be enforced in version 13.0.

The check on `amount_currency` is an extra check,
because if its different than the balance or not 0,
it might indicate this was a move line in a foreign currency
at some point, but somehow now the currency and the company currency are
the same. In such a case, there might be something else to correct.

closes odoo/upgrade#1405

Signed-off-by: Denis Ledoux (dle) <dle@odoo.com>