[FIX] mail@10.saas~15: missing create_column for is_mail_thread
is_mail_thread became a column of the ir_model table but a create_column
was missing. If the database tried to update ir_model records from the
new version, the upgrade would fail since it would try to set the
is_mail_thread column to something, but it is not created yet.

closes odoo/upgrade#1438

Signed-off-by: Christophe Simonis <chs@odoo.com>