[FIX] account@saas-12.3: make cash basis taxes migrate properly with credit notes
https://github.com/odoo/upgrade/commit/e0011923f3a0bd4e21c95445aef59f00473e73f1 aimed to fix cash basis taxes migration, but, just like the fix it was related to, it neglected the fact that a credit note and an invoice using cash basis taxes, when reconciled together, also generated a cash basis entry. This case is more complicated to handle, as such a cash basis entry doesn't give any unambiguous way of identifying what line comes from what document, which we need to know in order to apply the right sign multiplicator in the tax report.

We introduced the following fixes in stable, which change the behavior of the cash basis taxes, so that the tags are modified directly when creating the entries:
https://github.com/odoo/odoo/pull/53085
https://github.com/odoo/enterprise/pull/11193
https://github.com/odoo/odoo/pull/53503
https://github.com/odoo/enterprise/pull/11372

This way, the tax report doesn't have to do this mapping between cash basis lines and orginating documents ; instead, it always uses a multiplicator of 1 on top of the tag sign and the line's balance.

A sign inversion needs hence to be performed during migration on the tags of some cash basis tax lines. This commit should cover nearly all cases automatically, the only exception being the case in which an invoice with negative quantities is reconciled with a credit note: in this situation, we cannot easily distinguish between the lines coming from the invoice and the ones from the credit note. When it occurs, an entry is added to the migratino report, and some manual intervention will probably be required eventually.