[FIX] l10n_es_reports@saas~12.3: fix swapped XML IDs
In 12.0, two XML IDs were swapped (https://github.com/odoo/enterprise/commit/9aa7455).
If the database had them the old way and upgrades to saas~12.3, the upgrade will fail, as the upgrade scripts
will try to update the records pointed to by the XML IDs, but since they
are swapped it will create a record with a repeated key.

To fix this, the records in `ir_model_data` are assigned to the correct id
from `account_html_financial_report_line` during the upgrade to
saas~12.3.

closes odoo/upgrade#1456

Request: 49448
Signed-off-by: Christophe Simonis <chs@odoo.com>