[FIX] base@13.0: avoid crash on invalid VAT
Some databases contain invalid VATs (according to the check_vat method
introduced in the `base_vat` module), which causes an upgrade failure if
a the check is triggered during upgrade (this happens, for example, with
the installation of the `l10n_latam_base` module, which triggers it
within its install hook).

To avoid this, the definition of the check_vat method is changed during
the upgrade. But if this is done within the upgrade script for the
`base_vat` module, it will not be executed if this module is also
installed during the upgrade, which is the case in some of the databases
having this problem. Therefore, the definition of the change_vat model
is done in the upgrade script of `account`. If `base_vat` is installed,
the method will be overridden, and if it is not, the unneeded change_vat
method will be defined during upgrade, but it will be harmless.

Request ID: 45304

closes odoo/upgrade#1462

Signed-off-by: Denis Ledoux (dle) <dle@odoo.com>