[FIX] base@0.0.0: do not raise for deleting of mixin fields
There are a bunch of databases for which the upgrade fails
because of fields coming from mixins being deleted
which no longer exists in the database.

For instance, if you install `mass_mailing`,
it adds fields on `mail.compose.message`:
https://github.com/odoo/odoo/blob/c87db90d5dbab4d02a9378a1fc905c9b2935cbe2/addons/mass_mailing/wizard/mail_compose_message.py#L11-L14
`account.invoice.send` inherits from `mail.compose.message`:
https://github.com/odoo/odoo/blob/c87db90d5dbab4d02a9378a1fc905c9b2935cbe2/addons/account/wizard/account_invoice_send.py#L12

and therefore the fields `mass_mailing_id`, `campaign_id`, `mass_mailing_name`, `mailing_list_ids` are added to the model `account.invoice.send`
once `mass_mailing` installed, but because of a bug in the ORM, the fields are marked as belonging to the module `account`.

If you then uninstall `mass_mailing`, the `ir.model.fields` records remains, as it was `mass_mailing` which got uninstalled,
not `account`.

Then when the upgrade happens, the upgrade detects the fields should no longer be there (and this is correct)
and therefore the ORM decides to remove it.

The same thing happens with `alias_contact` of `crm.team` when `crm` gets uninstalled.
`crm.team` inherits of the `mail.alias` mixin:
```python
_inherit = ['mail.alias.mixin', 'crm.team']
```
There it's even weirder: the field is sometimes marked as belonging to the `hr` module:
```sql
select * from ir_model_data where model = 'ir.model.fields' and res_id = 2707;
  id   | create_uid | create_date | write_date | write_uid | noupdate |             name              |        date_init         |       date_update        | module |      model      | res_id
-------+------------+-------------+------------+-----------+----------+-------------------------------+--------------------------+--------------------------+--------+-----------------+--------
 24657 |            |             |            |           | f        | field_crm_team__alias_contact | 2019-07-23 01:24:57.3966 | 2019-07-23 01:24:57.3966 | hr     | ir.model.fields |   2707
(1 row)
```

This may comes from the fact `hr` overrides the selection of `alias_contact`:
https://github.com/odoo/odoo/blob/09dbbd77d83332c48f62b1c36c3b33ddbfa3a4eb/addons/hr/models/mail_alias.py#L10
And suddenly the ORM says this field comes from the module `hr`.

Request 49747

`It looks like you forgot to call `util.remove_field` on the following fields: account.invoice.send.mailing_list_ids`

closes odoo/upgrade#1474

Signed-off-by: Denis Ledoux (dle) <dle@odoo.com>