[FIX] product@saas~12.2: avoid computation of `can_image_be_zoomed`
Skip the computation of `can_image_be_zoomed` as it requires
the image in the filestore,
and the filestore is not there during upgrade.

Besides, it will be recomputed when the images will
be resized with the according cron of
`migrations/product/saas~12.2.1.2/resize-images-cron.xml`

Among others, this is to avoid the flood about
files not being in the filestore during an upgrade.
Also to gain some performances,
as it will no longer try to load files which are not
there during the upgrade :D.

I take the opportunity to solve other issues:
 1. Remove of the conditions `util.table_exists`
    it was util in the past because the script was imported in
    another script, in a module not depending on `product`.
    But today this import has been removed so its no longer useful
    5d7aa2c5a5ad7ea9ef14b964c3e1894e3ba3bf99
 2. `can_image_raw_be_zoomed` only exists on `product.product`,
    not on `product.template`:
    https://github.com/odoo/odoo/blob/2fa22c6f8bf844dc6e23a052c63af4ee231136d5/addons/product/models/product.py#L152
 3. `can_image_be_zoomed` is stored only on `product.template` (through the `image.mixin`),
    on `product.product`, it's a non-stored compute:
    https://github.com/odoo/odoo/blob/2fa22c6f8bf844dc6e23a052c63af4ee231136d5/addons/product/models/product.py#L161

closes odoo/upgrade#1477

Signed-off-by: Denis Ledoux (dle) <dle@odoo.com>