[FIX] account@saas~12.4: guess the move line account company is invalid
Guess the account to use when an invoice line uses an account which
is not in the invoice journal company.

Initially, the goal to set it to `NULL` was to let the ORM
choose the account itself,
but it doesn't always do the job correctly.

We therefore guess the account ourself,
using partially what is done in
`_get_computed_account`
which is the method called when the product is changed
on an invoice line,
in the method `_onchange_product_id`

Request 49577
M1809288652916

closes odoo/upgrade#1482

Signed-off-by: Denis Ledoux (dle) <dle@odoo.com>