[FIX] account@saas~12.4: `amount_total` of `entry` moves
The `amount_total` of moves of type `entry` not implying a receivable/payable account were set to 0

First, see the `_compute_amount` behavior:
https://github.com/odoo/odoo/blob/4f7981f80fb55ab4bf8b30d137e45b4fda4013cd/addons/account/models/account_move.py#L1054-L1077

For an `entry` move, the total amount is computed by summing
the balance of all debit lines of the move:
```python
if line.debit:
    total += line.balance
    total_currency += line.amount_currency
```
With the former condition of the upgrade script:
```sql
WHERE account_internal_type NOT IN ('receivable', 'payable')
```
If the move does not imply receivable/payable lines,
all lines of the move were taken into account, and,
as the debit lines have a positive balance,
as the credit lines have the opposite negative balance,
the resulting sum for the move was 0.

For instance, this is what happens in moves of assets:
no receivable/payable accounts are implied,
only assets/depreciation accounts.

This is specifically important to have this `amount_total` correctly
computed for the assets, as this is the field
used in the assets depreciation board to display the depreciation:
https://github.com/odoo/enterprise/blob/beee0f3690772ceadb588f45eaf8206c6b0294d9/account_asset/views/account_asset_views.xml#L97
Before this revision, the "depreciation" in the depreciation board of assets was always 0.

Even weirder: if there was receivable/payable lines,
but doing the counterpart of only part of the other lines,
the total was not 0, it was the amount of these payable/receivable lines only.

In a standard, normal, invoice, rejecting receivable/payables lines
is indeed a good strategy, as it will lead to only take into account
either all debit lines either all credit lines,
and therefore the correct total amount.

However, for a miscellaneous entry,
where receivable/payable lines are not there,
or do not counterpart all other lines, this is not enough.

The current condition nows reflects what `_compute_amount` does regarding entry moves.

Unit tests:
-----------

The upgrade of a miscellaneous assets move (`entry` move, no receivable/payable lines),
has been added to the already existing unit test testing the computation of the different invoice totals.

In addition, I take the opportunity to cover the case of the
 - sale refund invoice,
 - purchase invoice,
 - purchase refund invoice.

So we make sure the amounts are correctly computed for all types of invoices
(receipts are missing, but at least all combination of debit/credit on income/expense/receivable/payable accounts) are covered.

Revert of 92aea09f14e0d29af4e15447da30d3f0643060a2
--------------------------------------------------
The PR #1156 was attempting to solve this same issue,
but the fix was only partial,
as it was only correcting the `amount_total_signed`,
while the `amount_total` had to be fixed as well.

Besides, it was requiring an additional "heavy" SQL request.

Upgrade request 49772
M1612282607391

closes odoo/upgrade#1468

Signed-off-by: Denis Ledoux (dle) <dle@odoo.com>