[FIX] account_asset@saas~12.3: Manage asset moves grouped by category
Since odoo/enterprise@6cb107c42c1e3a679f7c7b2f41466f39d042bd16
this is no longer possible to group the accounting entries of
assets by categories.

Besides, the depreciation lines are now a one2many on `account.move`.
It therefore requires one `account.move` by depreciation line.

The strategy here is to split the move into pieces,
one by depreciation line the move contains.

The name of the resulting moves is the name of the initial move
with a counter (e.g. Move/001/1, Move/001/2, ...).
According to Quentin (qdp), this is acceptable, legally speaking.

The sum of debit, credit and balance of the resulting moves
must be equal to the initial move debit/credit/balance by account.
This is a requirement, legally speaking.

As this is a very delicate upgrade script,
as it could potentially mess up the accounting,
the upgrade will hard-fail if a case not implemented is detected:
 - The asset moves contain more than a debit and a credit lines:
   Normally, if the user do not mess up the move,
   there is by default only one debit and one credit line by asset move.
   If this is not the case, we raise, and we will then adapt the script
   so the upgrade can support the case,
 - The asset is in a foreign currency.
   Normally, the asset moves are always within the company currency.
   If this is not the case, we raise, and we will then adapt the script
   so the upgrade can support the case,
 - Once the move splitted into pieces, this is possible there is a difference
   compared to the initial move (e.g. rounding differences).
   In such a case, we add a last move containing the difference.
   However, if the difference is negative, we currently do not support
   the case as there was no example in the studied case, subject
   of this revision.

Request 49772
M1612282607391

closes odoo/upgrade#1483

Signed-off-by: Denis Ledoux (dle) <dle@odoo.com>