[FIX] event@saas~13.3: keep filter by tag active if enable before migration
event_tag_ctageory replace the old view event_tag.

It is not really a rename of view, but since the purpose is exactly the same
it make sense to keep the same state that previously.

So we rename it, and remove the flag no_update to be sure that arch is updated

Missing migration script to 197ae16 commit

closes odoo/upgrade#1489

Signed-off-by: Christophe Simonis <chs@odoo.com>