[FIX] product@saas~12.5: execute_values with page_size too small
Upgrade to saas~12.5 crashes in some databases with a lot of repeated
products for the same template.

To avoid having repeated products for the same template, the affected
upgrade script inserts generated technical attributes to these products
to make them unique.

The problem is with one of the `execute_values` calls which inserts
attribute values. This call is a `INSERT INTO...RETURNING`, because
after the insertion we put all the ids of the inserted records into a
dictionary, which is later iterated to generate a different query.

However, `execute_values` will execute the specified query in more than
one statement depending on the number of values passed, which is by
default 100 (see the documentation https://www.psycopg.org/docs/extras.html#psycopg2.extras.execute_values,
specifically the `page_size` arg). So if the values exceed 100, when we
fetch the results we will only get the ids of the last statement that
was executed. In the case of request 49968, the number of values is 104,
so the final fetch only got the last 4 elements.

To fix this, the `page_size` parameter is set to the maximum possible
number of values we can have, so only one statement is executed and the
`RETURNING` clause gives us all the ids inserted.

If we read the docs linked before, we see that another possible solution
would've been to use `fetch=True`, but apparently this parameter is not
available in the version of psycopg2 currently used by Odoo.

Request ID: 49968

closes odoo/upgrade#1499

Signed-off-by: Christophe Simonis <chs@odoo.com>