[IMP] account_asset: convert depreciation lines to account move using sql
For databases heavily using assets,
you might have to convert
multiple hundreds of thousands depreciation lines to accounting entries,
if you have a lot of assets with depreciation lines not yet posted
(in the future)

This just cannot be done swiftly using the ORM.

Therefore, convert the upgrade script
creating moves for asset depreciation line not having one yet
from Python to SQL.

For the case for which this was implemented
- 320.000 depreciation lines to convert to account moves -
this improves the processing time from a couple of days to less
than 5 minutes.

Besides, the current upgrade script doing the conversion
using Python doesn't handle the case of assets in foreign currencies
(the currency set on the asset is different than the company currency).
This is a valid case handled in standard 12.0/saas-12.3/13.0:

12.0: https://github.com/odoo/enterprise/blob/12.0/account_asset/models/account_asset.py#L537-L538
saas-12.3: https://github.com/odoo/enterprise/blob/saas-12.3/account_asset/models/account_move.py#L90
13.0: https://github.com/odoo/enterprise/blob/13.0/account_asset/models/account_move.py#L182

In such a case, the debit/credit on the move lines
must be the amount converted with the currency rate,
and the amount_currency must be set with the depreciation amount

This SQL upgrade script does it.
I could have done the bugfix for the Python version as well,
but as I intend to remove the python upgrade script
as soon as the SQL script is tested and validated for the studied case,
it's a bit pointless to do the bugfix of something which will be removed
in a couple of days.

A side note: There is a change of behavior regarding the currency
set on the move itself between 12.0 and saas-12.3
(on the move itself, not on the move lines):
- in 12.0, the move created for each line has the company currency,
  and the amount is therefore within the company currency
- from saas-12.3, the move created for each line has the asset currency,
  and the amount is therefore within the asset currency.
I chose to keep the behavior of 12.0 for the upgrade script,
because for assets having some depreciation lines posted,
and some not, you would otherwise end up
with assets having a depreciation board with some lines in the company currency
and some lines in the asset currency, and therefore a weird depreciation board
mixing the two currencies.

Request 44727
M1503061860266

closes odoo/upgrade#1523

Signed-off-by: Denis Ledoux (dle) <dle@odoo.com>