[FIX] purchase_stock@saas~12.4: Fix amount_untaxed with anglo saxon lines
If you make a purchase order to buy a product with a cost of 9 but you make the related vendor bill buying at a cost of 12, an extra move line of 3 will be generated on the price difference account.
At this point of the migration, your account.move.line has price_unit=12 and credit=9 meaning the amount_untaxed of the vendor bill is now wrong: amount_untaxed=9 because the total is based on the accounting values.
Also, the anglo saxon lines are removed when the vendor bill is reset to draft meaning Odoo will attempt to unlink the original invoice line. In order to fix the problem, we need to create an additional journal item containing the amount corresponding to the price difference.

closes odoo/upgrade#1530

Signed-off-by: Nicolas Seinlet (nse) <nse@odoo.com>