[FIX] base@0.0.0: transient models without table
The upgrade scripts delete all data from tables associated to
transient models. However, it may be the case that there is a model that
has no table associated, due to some error (obsolete model from a past
version not being correctly deleted) or by customer action. In these
cases, the upgrade fails.

To fix this, we check if the table associated to a transient model
exists before trying to delete its data.

Request ID: 51716

closes odoo/upgrade#1542

Signed-off-by: Christophe Simonis <chs@odoo.com>