[IMP] hr_holiday@0.0.0: rename calendar.events
Before odoo/odoo@ccfe9f3bd72ab48bb40176cfc501c5566bb8f515, the
calendar.events that were created from the hr.leaves contained the type
of leave, which is considered as a confidential information.

In order to remove this information from the calendar, these old events
need to be renamed to the new format, which does not include the type of
leave any more.

closes odoo/upgrade#1545

Signed-off-by: Christophe Simonis <chs@odoo.com>