[FIX] account_asset@saas~12.3: amounts with negative values
Some databases have negative values in the `amount` field in the
`account.asset.depreciation.line` model. These negative values are the
result of an old bug probably caused by a rounding error. An asset that
should be depreciated linearly until the value is 0, gets brought down to slightly
below zero (because of rounding of the linear factor). This results in a
negative `remaining_value`, so the next depreciation has a negative
amount to bounce it back to 0 (usually small, less than 1$).

Before saas~12.3, a single move could group many depreciations of an
asset. During the upgrade to saas~12.3, a separate move is created for
each depreciation line. But it is not possible to create a move with a
negative amount, so the upgrade fails when it encounters the negative
amount in these databases.

To fix this, when we encounter a negative amount in a depreciation line,
we invert the debit and credit values, which is what would happen in
Odoo if we have a negative depreciation amount intentionally.

The upgrade scripts also check that the total amount of the moves before
and after the upgrade are consistent. However, they will not be
consistent if there were negative amounts. This is because the grouped
move was the sum of all amounts in the depreciation, including
negatives. But when we create move lines for the negative amounts, we
use positive values, we just invert the debit and credit accounts.
Therefore, in this case we will only check that the balance is
consistent, which should never change.

Request ID: 45448

closes odoo/upgrade#1570

Signed-off-by: Denis Ledoux (dle) <dle@odoo.com>