[FIX] hr_holidays@12.5: make Everyone menu visible by anybody
hr_holidays.menu_hr_holidays_dashboard was formerly called My Team, and
therefore was only visible by the group Time Off / Administrator.

Since odoo/odoo@e0b16b22e8a1e6cdf55a8656042db5ae356de15d, this menu was
renamed to Everyone and is not limited to a group any more. It was
probably a mistake to reuse the same external identifier, as the groups
are kept during a migration because the groups attribute is not
specified any more.

This PR fixes the issue by removing any group from the menu item.

closes odoo/upgrade#1552

Signed-off-by: Christophe Simonis <chs@odoo.com>