[FIX] testing: allow commit for log_db
Integrity case forbid commit to avoid to accidentally modify a database
when executing tests.

Unfortunately, the generic patch was also applied on log_db cursor. Errors and
warnings where not displayed on runbot forcing users to access text logs.

This commit allow the cursor to commit if the corresponding database is the
log_db. Using the connection to commit is ok since this transaction does not
use any cache.

closes odoo/upgrade#1554

Signed-off-by: Christophe Simonis <chs@odoo.com>