[FIX] account@9.0: allow pre-computation of reconcilation fields
The purpose is to be able to pre-compute the fields
`amount_residual`
`amount_residual_currency`
`reconciled`
of `account.move.line`
(To put them in database before the upgrade happens)
as their computation takes a while, especially for
databases heavily using accounting.

The script triggering the computation only for the move lines with
these values to `NULL` is already there:
https://github.com/odoo/upgrade/blob/a023c105a4bc3ecf5bba0e36c4529ceccf40656e/migrations/account/9.0.1.1/post-90-compute-field.py#L37

In the SQL statement modified by this revision,
replacing `NULL` by the current value of the table:
 - will have no effect for databases not pre-computing the values,
   as the columns has just been created a few lines above, without default set,
   and therefore the value for these columns are already NULL
 - will allow to pre-compute the value for willing databases,
   as it will recover the value which was already computed
   in advance.

Previously, in 8.0, `amount_residual` and `amount_residual_currency`
were computed but not stored, and therefore the column should not be
there.

If a database did a customization to store these fields,
then they will already hold the expected value, so no worries either.

Request 45798
M120530613-xpixzyur

closes odoo/upgrade#1576

Signed-off-by: Denis Ledoux (dle) <dle@odoo.com>