[FIX] product@saas~12.2: handle resize of invalid/incomplete images
When doing an upgrade with the filestore included,
-
which doesn't happen a lot, since in most cases the filestore is not
provided, and even if provided, the current upgrade platform discards
it before the upgrade. For the current case,
it's because the database upgrades from 8.0 to 13.0,
and the files were in database and are converted to filestore
during the upgrade.
-
the images are being resized during the upgrade.
However, if the image is corrupted/invalid,
the resize will fail and this will prevent the upgrade.

Since we have added a few revision ago a cron
meant to resize images that were not resized
during the upgrade,
when such an image resize failure occurs,
we just had it to the images to be resized by the cron/after upgrade.
So it re-attempts to resize the image later.

Request 45798
M120530613-xpixzyur

closes odoo/upgrade#1578

Signed-off-by: Denis Ledoux (dle) <dle@odoo.com>