[FIX] account@8.saas~6: deleted sequences being updated
The model `account.sequence.fiscalyear` has a main sequence and a
sub-sequence. Before this model is deleted in the upgrade scripts, the
main sequences are updated, and the corresponding sub-sequence is
deleted and all references to it are changed to the main sequence.

However, in some databases it is the case that the sub-sequence in a
record also appears as a main sequence in another, so the upgrade fails
because it tries to update a deleted record.

To solve this, we keep track of the sub-sequences we have already treated, and
if we find its id as a main sequence later we just skip it.

Request ID: 46548

closes odoo/upgrade#1579

Opw: 2242719
Signed-off-by: Daniel Monzonís <dmonzonis@users.noreply.github.com>