[FIX] mail@saas~11.1: circular reference in mail_test
There is a circular reference between the `mail.test` and `mail.alias`
models which causes an error during the upgrade. The table `mail_test`
has the id of a `mail_alias` as a foreign key, with constraint set to
`ON DELETE RESTRICT`, and `mail_alias` has an indirect reference to
`mail_test` via its `alias_model_id` column.

During the upgrade to `saas~11.1`, the `mail.test` model is removed via
the `remove_model` method, which first removes all references to it. But
when it tries to remove the records in `mail_alias`, it fails due to the
constraint in `mail_test`.

To fix this, before removing the model, we remove the constraint so that
we can remove the `mail_alias` records before dropping the `mail_test`
table. We do this instead of deleting the records in `mail_test` first
in case there is another reference to them in some other table that
needs to be removed too.

Request ID: 45886

closes odoo/upgrade#1603

Signed-off-by: Nicolas Seinlet (nse) <nse@odoo.com>