[FIX] stock@tests: products with mixed bom types not being ignored
The integrity test `TestOnHandQuantityUnchanged` checks that no
product quantities have changed during the upgrade. Ignoring
products that are associated to bom kits when upgrading to 13.0 or after
since the computation of the available quantity for bom kits changed
between the versions.

However, if a product is associated to more than one bom and has mixed
types (some boms are phantom and some normal), as long as the product is
associated to at least one phantom type, the available quantity will be
computed using the algorithm for bom kits. This case was not correctly
considered in the tests, so it would not ignore products with mixed bom
types, resulting in the test sometimes failing due to these products
having different quantities before and after upgrade.

Request ID: 45886

closes odoo/upgrade#1621

Signed-off-by: Denis Ledoux (dle) <dle@odoo.com>