[FIX] account_asset@saas~12.3: remaining moves with negative amounts
In commit
https://github.com/odoo/upgrade/commit/2b8cf6da58d86af31303b1bf22685d88d79ea13e,
the case where negative amounts exist in an asset depreciation line was
treated, by creating a move with the positive value but with the debit
and credit switched. This ensures that the balance remains constant,
however, there can be a difference between the previous move amount and
the sum of the amounts of the newly created moves (since the amount is
always positive).

But during the upgrade script, we check if there is a difference in the
amounts and create another move if some difference exists. Since the
move amounts may vary if there were negative amounts in the depreciation
line, this will not work in all cases. Instead, we should check that the
total balance of the move lines for each move remains constant.

This commit does just that. We look for differences in the total balance
of each move line between the new moves created and the old grouped
move, which could be caused by rounding errors among other things. For
each of the differences, we create a move with move lines that make up
for the difference.

Request ID: 52600

closes odoo/upgrade#1637

Signed-off-by: Denis Ledoux (dle) <dle@odoo.com>